refactor: move response utils
