+++ server/server.js
@@ -1,6 +1,7 @@
 import { Sequelize } from 'sequelize';
 import { fingerprintJsApiClient } from './fingerprint-api';
 import { CheckResult } from './checkResult';
+import { sendForbiddenResponse } from './response';
 
 // Provision the database.
 // In the Stackblitz environment, this db is stored locally in your browser.
@@ -67,38 +68,6 @@ export async function getVisitorDataWithRequestId(visitorId, requestId) {
   return fingerprintJsApiClient.getVisitorHistory(visitorId, {
     request_id: requestId,
   });
-}
-
-/**
- * @param {import("next").NextApiResponse} res
- * @param {import("./CheckResult").CheckResult} result
- * */
-export function sendOkResponse(res, result) {
-  if (res.headersSent) {
-    console.warn('Attempted to send a OK response after headers were sent.', {
-      result,
-    });
-
-    return;
-  }
-
-  return res.status(200).json(result.toJsonResponse());
-}
-
-/**
- * @param {import("next").NextApiResponse} res
- * @param {import("./CheckResult").CheckResult} result
- * */
-export function sendForbiddenResponse(res, result) {
-  if (res.headersSent) {
-    console.warn('Attempted to send a forbidden response after headers were sent.', {
-      result,
-    });
-
-    return;
-  }
-
-  return res.status(403).json(result.toJsonResponse());
 }
 
 // Report suspicious user activity according to internal processes here.
